+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 ## [0.1.1] - 
 
 - Rework authorization module 
-
+- Add single-qubit entropy
+  
 ## [0.1.0] - 2025-09 -15
 
 - First private release